works but need to modify the vector

+++ fdf.h
@@ -6,7 +6,7 @@
 /*   By: rkrief <rkrief@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/08 16:10:34 by rkrief            #+#    #+#             */
-/*   Updated: 2018/02/12 18:26:38 by rkrief           ###   ########.fr       */
+/*   Updated: 2018/02/15 14:20:39 by rkrief           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,16 +43,16 @@ typedef struct	t_pos
 
 typedef struct 	t_new
 {
-	int			x;
-	int			y;
-	int			z;
+	float		x;
+	float		y;
+	float		z;
 }				s_new;
 
 typedef struct t_cam
 {
-	int			x;
-	int			y;
-	int			z;
+	float		x;
+	float		y;
+	float		z;
 }				s_cam;
 
 int		**ft_parsing(char *str, s_check *numb);
